Documentation + namespace change + restructuralization
