+++ Recommendations.SDK/Kentico.Kontent.Recommendations.csproj
@@ -2,16 +2,16 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <RootNamespace>Kentico.Kontent.Recommender</RootNamespace>
+    <RootNamespace>Kentico.Kontent.Recommendations</RootNamespace>
     <Description></Description>
     <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput)</TargetsForTfmSpecificBuildOutput>
-    <Copyright>(c) 2018 Kentico Software. All rights reserved.</Copyright>
-    <AssemblyTitle>Kentico Kontent Recommender AspNetCore SDK</AssemblyTitle>
-    <Title>Kentico Kontent Recommender Netstandard SDK</Title>
+    <Copyright>(c) 2020 Kentico Software. All rights reserved.</Copyright>
+    <AssemblyTitle>Kentico Kontent Recommendations SDK</AssemblyTitle>
+    <Title>Kentico Kontent Recommendations SDK</Title>
     <NeutralLanguage>en-us</NeutralLanguage>
     <Authors>Kentico Software</Authors>
-    <AssemblyName>Kentico.Kontent.Recommender</AssemblyName>
-    <PackageId>Kentico.Kontent.Recommender</PackageId>
+    <AssemblyName>Kentico.Kontent.Recommendations</AssemblyName>
+    <PackageId>Kentico.Kontent.Recommendations</PackageId>
     <PackageTags>kentico;kontent;recommendations</PackageTags>
     <PackageIconUrl>https://github.com/Kentico/Home/blob/master/images/kk-logo-nuget.png?raw=true</PackageIconUrl>
     <PackageProjectUrl>https://github.com/Kentico/recommender-sdk-net</PackageProjectUrl>
