+++ Recommendations.SDK/Models/RecommendationSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Kentico.Kontent.Recommender
+﻿namespace Kentico.Kontent.Recommendations.Models
 {
     /// <summary>
     /// Extra recommendation options
