--- Recommender.SDK.CookieHelpers.MVC/Kentico.Kontent.Recommender.SDK.CookieHelpers.MVC.csproj
@@ -1,49 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net461</TargetFrameworks>
-    <RootNamespace>Kentico.Kontent.Recommender.MVC</RootNamespace>
-    <Description></Description>
-    <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput);IncludeP2PAssets</TargetsForTfmSpecificBuildOutput>
-    <Copyright>(c) 2018 Kentico Software. All rights reserved.</Copyright>
-    <AssemblyTitle>Kentico Kontent Recommender SDK for .NET MVC</AssemblyTitle>
-    <Title>Kentico Kontent Recommender SDK for .NET MVC</Title>
-    <NeutralLanguage>en-us</NeutralLanguage>
-    <Authors>Kentico Software</Authors>
-    <AssemblyName>Kentico.Kontent.Recommender.MVC</AssemblyName>
-    <PackageId>Kentico.Kontent.Recommender.MVC</PackageId>
-    <PackageTags>kentico;mvc</PackageTags>
-    <PackageIconUrl>https://github.com/Kentico/Home/blob/master/images/kk-logo-nuget.png?raw=true</PackageIconUrl>
-    <PackageProjectUrl>https://github.com/Kentico/recommender-sdk-net</PackageProjectUrl>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>https://github.com/Kentico/recommender-sdk-net.git</RepositoryUrl>
-    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="JWT" Version="5.3.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="Microsoft.Sourcelink.Github" Version="1.0.0" PrivateAssets="All" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Recommender.SDK.Shared\Kentico.Kontent.Recommender.SDK.csproj" PrivateAssets="All" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Web" />
-  </ItemGroup>
-
-  <Target Name="IncludeP2PAssets">
-    <ItemGroup>
-      <BuildOutputInPackage Include="$(OutputPath)Kentico.Kontent.Recommender.Shared.dll" />
-    </ItemGroup>
-  </Target>
-</Project>
