--- Recommender.SDK.CookieHelpers/VisitDetail.cs
@@ -1,8 +1,0 @@
-﻿namespace Kentico.Kontent.Recommender.SDK.CookieHelper
-{
-    public class VisitDetail
-    {
-        public string Ip { get; set; }
-        public string Referer { get; set; }
-    }
-}
