+++ Recommender.SDK.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30204.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.Recommender.SDK.CookieHelpers", "Recommender.SDK.CookieHelpers\Kentico.Kontent.Recommender.SDK.CookieHelpers.csproj", "{3305E2D2-C398-4D56-9E3B-1BE299502FE3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.Recommendations.CookieHelper", "Recommendations.SDK.CookieHelper\Kentico.Kontent.Recommendations.CookieHelper.csproj", "{3305E2D2-C398-4D56-9E3B-1BE299502FE3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.Recommender.SDK", "Recommender.SDK\Kentico.Kontent.Recommender.SDK.csproj", "{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.Recommender.SDK.CookieHelpers.MVC", "Recommender.SDK.CookieHelpers.MVC\Kentico.Kontent.Recommender.SDK.CookieHelpers.MVC.csproj", "{6B13C145-C46E-4A37-8B82-B2906B1FDE16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.Recommendations", "Recommendations.SDK\Kentico.Kontent.Recommendations.csproj", "{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +21,6 @@ Global
 		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C7BF0DCE-6004-4B3A-90BC-9D8C6E99128F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6B13C145-C46E-4A37-8B82-B2906B1FDE16}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
